fix: Fix of issue where the same images was downloaded even when requested new ones
When requesting images from Flickr api  (flickr.photos.search)
with a page parameter above around 200, it looks like Flickr is returning
the same set of images regardless of the pagenumber requested.
This commit is submitted to avoid this.
However, it will then limit the amount of possible photos
to be downloaded to 6000 (30*200)